LeNet/requirements.txt
numpy == 1.17.4
tensorflow-gpu == 1.13.1